Make CI green: 23 real type errors fixed, 40 that were never ours
ci.yml has been failing on two of its five checks, and neither was noticed
because the handoff's working standard is not the workflow's. "Keep ruff
green and mypy at 63" passes `ruff check` and treats a count as a baseline;
CI runs `ruff format --check` as a separate step and `mypy viflap` expecting
zero. A session following the handoff faithfully still pushed red.

Formatting: all seventeen files swept. reserved.py was among them, so the
speaker set was captured before and re-imported after -- 100 speakers,
identical. Reflowing a frozenset literal cannot change its membership, but
that file is a protocol decision and deserves the evidence rather than the
confidence.

Of the 63 mypy errors, 23 were real:

  - Six int(payload[...]) sites in pipeline.py carried
    `# type: ignore[arg-type]` where the error raised is call-overload, so the
    ignores suppressed nothing and were themselves reported as unused.
  - voice_quality.py declared it returned NDArray[int64] glottal marks. It
    builds float64 and always has -- the marks are sub-sample peak positions
    and `current` is constructed with float(). The annotation was never true.
  - InMemoryUnitOfWork did not satisfy the UnitOfWork protocol. A protocol's
    mutable attributes are invariant, so exposing the concrete repository
    types broke the match. The concrete references are now private, which is
    where they belong: staging is not part of the port.
  - bootstrap.py imported EvidenceBundle from the wrong module, and its
    unfitted _score stub returned a type incompatible with its supertype.
  - dependencies.py returned Any out of FastAPI's untyped app.state.

The other 40 were numpy, not us. Under numpy 2.x, NDArray[np.float64] is
ndarray[tuple[Any, ...], dtype[float64]], so .shape is tuple[Any, ...] and
.shape[i] is Any; several reductions erase the same way. Ten lines with
entirely correct annotations reproduce it, and that repro is in the config
beside the override so the next person does not have to rediscover it.

So warn_return_any is disabled for the twelve numpy-backed modules and stays
on everywhere else. The list is explicit rather than a wildcard, so a new
file has to be added deliberately instead of inheriting an exemption. The
alternative was forty casts threaded through plda, gmm, ivector and pav --
the code every reported number comes out of -- each asserting exactly what
the annotation above it already says. Revisit when numpy's stubs carry
concrete shape types.

832 tests pass. No runtime behaviour changes: every edit is an annotation, a
cast, a deleted dead ignore, or a rename of a private attribute.

+++ viflap/analysis/dsp/voice_quality.py
@@ -203,7 +203,7 @@ def measure_voice_quality(
 
 def _place_period_marks(
     signal: NDArray[np.float64], f0_track: F0Track, voiced: NDArray[np.bool_]
-) -> list[tuple[NDArray[np.int64], NDArray[np.float64]]]:
+) -> list[tuple[NDArray[np.float64], NDArray[np.float64]]]:
     """Locate one amplitude peak per glottal period, grouped into voiced runs.
 
     Marks are placed by **waveform cross-correlation**, pitch-synchronously.
